+++ packages/content-builder/src/resources/views/categories.blade.php
@@ -169,7 +169,6 @@
 
         function drawTable($data){
 
-
             $html = "";
             $.each($data, function(i) {
